sportybot/requirements.txt
pyTelegramBotAPI==4.18.0
google-genai>=1.0.0
requests==2.32.3
python-dotenv==1.2.1
tenacity==8.3.0
fake-useragent==2.0.3
beautifulsoup4==4.12.3
httpx>=0.28.1
